# openapi/blackglass.yaml
openapi: 3.1.0
info:
  title: BLACKGLASS API
  version: 0.1.0
  description: >
    Operational integrity API — hosts, baselines, scans, drift findings,
    evidence bundles, and reports. Idempotent reads; scan jobs are asynchronous.
servers:
  - url: https://api.blackglass.invalid/v1
    description: Production (stub host)
  - url: https://console.blackglass.invalid/api/v1
    description: SaaS console API prefix (Next.js `/api/v1`, Clerk session)
...
paths:
...
  /reports:
    post:
      tags: [Reports]
      summary: Generate integrity report
      security:
        - SessionCookie: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                fleet_tags: { type: array, items: { type: string } }
                period: { type: string, description: ISO-8601 interval or preset }
                format: { type: string, enum: [markdown, pdf], default: pdf }
      responses:
        "202":
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id: { type: string }
                  status: { type: string, enum: [queued, running, ready, failed] }
        "401": { description: Unauthenticated }
        "403": { description: Forbidden }
        "429": { description: Rate limit exceeded }
    get:
      tags: [Reports]
      summary: List generated reports
      security:
        - SessionCookie: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        title: { type: string }
                        format: { type: string, enum: [markdown, pdf] }
                        status: { type: string, enum: [queued, running, ready, failed] }
                        created_at: { type: string, format: date-time }
                  next_cursor: { type: string, nullable: true }
        "401": { description: Unauthenticated }
        "403": { description: Forbidden }
...
components:
...
  securitySchemes:
    BearerIngestKey:
      type: http
      scheme: bearer
      description: INGEST_API_KEY bearer token for push-agent endpoints
    SessionCookie:
      type: apiKey
      in: cookie
      name: bg-session
      description: >
        Signed session cookie (legacy auth) or Clerk session token
        (when CLERK_AUTH_ENABLED=true). All authenticated console API
        paths require one of these. Clerk issues __session cookie;
        the server inspects CLERK_AUTH_ENABLED to pick the verification path.
...